Update patch set 1
Patch Set 1:

I like the change, however I'm surprised that the build system would pick the default Java installation: starting in Gingerbread, it is supposed to explicitly pick the proper version of Java. The goal was to have Java 5 as default (for Froyo and earlier, which required Java 5 to be the default), while Gingerbread and later would ignore the default and pick their own.

Patch-set: 1
Label: Code-Review=0