Update patch set 1
Patch Set 1:

That'd be good. I'm worried that hinting too strongly at making Sun JDK 6 the default would get in the way of either building froyo (which expects Sun JDK 5 as default) or other Ubuntu software (which might expect OpenJDK).

Feel free to leave notes about the exact in the "known issues" section, specifically with the kind of error message you got and the workarounds you used.

Thanks for being patient, for investigating the situation, and for working with us to make sure that other people don't hit the same issue.

Patch-set: 1