Update Mac build environment documentation
This breaks the build environment instructions into branch-specific sections so that the documented environment requirements for each branch may evolve independently. Since I have only tested builds on 10.6 and 10.7, I tried to leave all other sections as untouched as possible.

While I cannot say that the instructions in the "android-4.0.3_r1 branch and all earlier branches" section are exactly correct, this change makes it obvious that building Master on a modern OS X environment is extremely easy. I think this will be valuable to developers who choose to work on Macs.

I'm trusting that developers who actually need to build earlier branches will update the documentation in those sections.

A live version of this change is available here:

http://mohsdev.com:8080/source/initializing.html

Change-Id: Ib41510f80c404f74e49e59d49e93f513d2ed8507