Update patch set 2
Patch Set 2:

Based on the way things have worked in the past, it's much more likely that a minor release wouldn't change the situation.

That way, with the current description, we have at most one change to do (in case there's a fix in a future 4.0.x), whereas if we specified an explicit range we'd have to update it for each release that doesn't fix things.

Patch-set: 2
Label: Code-Review=0