Update patch set 5
Patch Set 5:

Xcode 4 can also be obtained via the Mac App Store on Lion and the latest Snow Leopard without the need for a developer account. Is it worth updating the text to reflect that or would it be better for me to submit another patch once this has gone through?

Patch-set: 5
Reviewer: Gerrit User 1004614 <1004614@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0