Update patch set 3
Patch Set 3:

Could we get some clarification on some of the comments/questions in the commit message?  e.g. Do we need to keep this text applicable to 4.4 for now (if so, then we can't recommend use of permissive_or_unconfined yet) or will this not go live until after the next release?  And do we need to keep the URLs and comments about cloning/copying external/sepolicy given that it is already part of the AOSP source tree ever since 4.3?  At least for kernel/common, I understand since it isn't normally part of a repo client.
Then, lastly, for further changes, should we do them on top of this one or continue updating this change in place?

Patch-set: 3
Reviewer: Gerrit User 1010099 <1010099@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0