Update patch set 3
Patch Set 3:

My understanding, from having chatted with Clay in the past, is that documentation updates submitted through AOSP can go live at any time, and will *not* be tied to an Android release. Submissions through AOSP are considered public and available to be published as soon as practical.

If you want these changes to be tied to an Android release, I'd cherrypick your change and submit it to our internal tree. The internal tree would only be published when a new release of Android goes out the door.

For things like permissive_or_unconfined() vs permissive, IMHO, it would be best to structure the documentation to discuss both. Something like: "In Android 4.4, per-domain permissive is controlled by the permissive statement. In later versions of Android, per-domain permissive is controlled by the permissive_or_unconfined() macro" or soemthing like that.

I would remove the statements about cloning/copying external/sepolicy. In retrospect, it just doesn't make a lot of sense.

As for further work: Personally, I prefer lots of small changes showing the revisions over time. I'm not sure if Clay has a different preference.

It might be easier to start a Google doc with proposed changes / rewrite, iterate on that document, and then cut/paste the final version into a .jd file. Google docs is a more user friendly way of seeing/editing/commenting on large documentation changes. 

Clay: do you have a preference?

Patch-set: 3
Label: Verified=0