Update patch set 2
Patch Set 2: Code-Review-2

Hey Clay - This needs to be more generic. The security levels are specific to Widevine and others may have other levels. What is needed is to detail the framework / plugin arch, facilities for developing plugin and security. Lets sync when I am back in office next week. Sorry if I was not crisp earlier

Patch-set: 2
Label: Code-Review=-2