SELinux index:  Improve documentation.
Reorder supporting documentation links to prefer the most specific
and up-to-date resources over the most general or older resources.

Add a link to the SELinux Object Classes and Permissions Reference
page.

Add a link to the Configuring the SELinux Policy technical report
as it was written in the same timeframe as the already linked
Implementing SELinux report and describes the policy concepts and language
as they existed at the time.  However, we could perhaps dispense
with both links as neither is particularly up-to-date.

Change-Id: I623d6ed5a8b8a98a9fab8f8d6eb318eab19d798e
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>