Update patch set 2
Patch Set 2:

> (1 comment)

I don't know the exact reason to pack images into the target files zip. But probably to make block-based OTA generation easier.

And for most of the cases, they should be identical for the ones in the signed target files zip and the separate img.zip. Unless some options are specified when generating the signed files.

Patch-set: 2
Label: Code-Review=0