Update patch set 6
Patch Set 6: Code-Review+2

> (2 comments)
 > 
 > Hi Elliott,
 > 
 > I've given this page a new title. I can also use "Native Debugging
 > of the Android Platform"

it's more like "Debugging Native Android Platform Code" (as opposed to Java code). really, i think this is a sign that we shouldn't be sneaking this stuff onto the main page and should just have an explicit "Native Debugging" page (and maybe a "Java Debugging" page if you can find someone with something to say; but maybe if we clear out the top page *that* would then be a good place to say that Java debugging should "just work" from your chosen IDE).

whatever, this is definitely an improvement on what's there today, so +2.

Patch-set: 6
Label: Code-Review=+2