Update patch set 2
Patch Set 2:

As far as I understand, cygwin is not a virtual machine, but rather a dll which translates posix code to the windows equivalent.  Thus, the ABI would change and any native code would need to be recompiled.  In short, we can't run native code on windows hosts, even with cygwin.

Patch-set: 2