Update patch set 8
Patch Set 8:

Can you modify the table so the languages and patch string don't span multiple lines?  If needed, can we take some space from the first column?  With the line break, it looks odd.

For the languages, is zh-CN, zh-TW common or should we use a proper string?  I know it'll make the table look bigger but I'm wondering if the shortform is understandable by someone not familiar with the shortnames of those languages.

Patch-set: 8