CDD: Clarify conditions when Vulkan native APIs can be used Section: 3.3.1. Application Binary Interfaces
The use of the Vulkan native library name was reserved as the
specification was not yet launched at time of when Android 6.0
shipped.

As now the Vulkan 1.0 is launched, we're amending the CDD to
conditionally allow the use of the Vulkan native library.

See https://www.khronos.org/vulkan/ for more detail.

Bug: 25931918
Change-Id: Ic5132aa3c47c6fc9b8a7c7fbd35c15f068f2162d