Update patch set 8
Patch Set 8:

Sorry, I'm not following the history well here, I didn't think we generally wanted to have directions for non LTS releases, as that is not something we use internally? I leave it to Neil I guess. I'd think once 16.04 LTS comes out in April, we would want to just replace 15.x references with that and leave the non-LTS releases as an exercise for the reader.

Patch-set: 8
Label: Verified=0