CDD: Clarify fingerprint false rejection rate requirement Section: 7.3.10. Fingerprint Sensor
Multiple inquires have been made on whether the false rejection
rate requirement describes the specification the fingerprint sensor
component or the end-to-end results as measured on the device.

The false rejection rate of the later could be measured at a higher
level because of factors like ergonomics etc., so hereby the STRONGLY
RECOMMENDED false rejection rate requirement is clarified to be
applied at the device implementation level.

Bug: 25015187
Change-Id: I667cd5bf2b6c6fff820afe8e6a1fc273f0c624ca