CDD: SD card encryption required for the primary external storage Section: 9.5. Multi-User Support
The CDD language made sense until Android 4.3, where the only
"external storage APIs" were for the primary external storage.

However with Android 4.4 the definition of external storage APIs
became ambiguous as apps gained write access to app-specific
directories on secondary external storage devices.

The AOSP implementation of these APIs only offers write access
to secondary storage for the owner user, so it is worth to clarify
that the encryption of the removable storage media for Multi-User
support is only required if the removable storage media is used
as the primary external (a.k.a. application shared) storage.

Bug: 22699648
Change-Id: Ie8568272eec05392998bcfe398e82ae9347fcb9c