CDD: Allow the change of screen density changes Section: 7.1.1.3. Screen Density
Due to the rise of phablets, and the evolution of
screen technologies, it is more relavant to visit
the possibility to find a path to allow the screen
density to be customized to the user's needs while
minimizing the impact to apps.

This change introduces a controlled way to allow
user-configurable screen density changes of the
default display, while providing the mechanism
to mitigate challenges pre-installed app distribution
services has to deal with.

Bug: 27669029

Change-Id: Ifa501dd65a2b2491d9afd3b20974e10f4cee7168