CDD: Clarify that UHD HEVC decoders requires Main10 Level 5 Main Tier profile Section: 5.3. Video Decoding
The original CDD language calls out the requirement to support
Main10 Level 5 Main Tier profile, which supports 10-bit depth,
together with the recommended requirement to support the UHD
decoding profile.

As such, it may be misunderstood as that when implementing the UHD
decoding profile, the support of 10-bit color depth is a separate
and an additional recommendation and not required.

However as being validated in the following CTS test,
the intent is to always require Main10 Level 5 Main Tier
profile when UHD (4K) HEVC (H.265) decoding is supported.

MediaCodecCapabilitiesTest#testH265DecoderProfileAndLevel

Bug: 25744892
Change-Id: Ie9cf34614e37e31e964ead6c6bf370d9cb4112c3