Update patch set 2
Patch Set 2:

Now that the "6.0" is hardcoded into these files, the HTML file should be android-cdd-6.0.html.  This also distinguishes it from the old android-cdd.html, which was the working file. Whether or not we should hard code the Android version in the jd filename is a different thing--I could see us going either way with that one.

Patch-set: 2