Update patch set 2
Patch Set 2:

But that would mean needing to update links in body and navigation every time you push a new version. This is why I removed the version name from the PDF long ago, to avoid that maintenance and potential breakage.

Since we're building the HTML in a separate project, why not check it in there and the modified .jd here? Thanks.

Patch-set: 2