Update patch set 3
Patch Set 3:

(2 comments)

Nice effort! 

i) Can we keep the src/compatibility/android-cdd.html doc "as-is" it is like the 'master' branch -- and the version specific html files are the final (and more up-to-date) ones.
ii) android-cdd.jd is based on android-cdd.html which is not necessarily up-to-date with the latest 6.0 CDD
iii) I can be persuaded either way, but not sure if we want to add a new page or just have the table added to the overview page
iv) the "download" page also have links to the CDD, probably we can also remove all links from there but instead add a link from that page to the page that will have the CDD table

Patch-set: 3