CDD: Clarify TV app req. to display both installed and 3p-app inputs
The SDK document (http://developer.android.com/training/tv/tif/tvinput.html)
describes that a TV input service represents a media stream source, and that
implementing it allows the app developers' media content to be presented in
"a linear, broadcast TV fashion as channels and programs".

Thus it sets the expectation for app developers that their TV inputs would
be shown in the default TV experience along the other media stream sources
such as a TV tuner that is pre-installed on the device.

While there is already a CTS verifier test in Android CTS 6.0 validating
this behavior of the TV Input service APIs documented in the SDK, it seems
there is value to more explicitly document this in the CDD.

Bug: 25487122
Change-Id: Ie55a944aaff5443d54bb2fc23b1c1c89e57e0c5c