Update patch set 4
Patch Set 4:

> I asked for the code formatting as I thought stack was a command
 > line tool. Incorrect?

there is such a tool, yes, but that word "stack" in that particular sentence refers to the logcat output. the "stack tool" in the same sentence refers to <code>stack</code>. (but like i said, better than code font would be a link to the docs we're about to add for that tool.)

(this would be less confusing if the tool had been called "decode-stack" or something! nouns make bad command names!)

Patch-set: 4