Update patch set 1
Patch Set 1:

The CLs listed here are using the "brief" summary. There is also the option to include the full commit message, but I found that was leaking too much info. If we want, we can switch back to that after the initial copybara sync and we're in "iterative" mode (still not sure what that looks like).

The CLs are autogenerated, I don't have a whole lot of say beyond the settings that copybara provides. Copybara created this CL directly and didn't use any of my working branches, so I'll have to look into how to take over the CL and amend the commit message. I'll need to find out if modifying/removing entries affects copybara's ability to sync history in future commits.

Patch-set: 1
Reviewer: Gerrit User 1125758 <1125758@85c56323-6fa9-3386-8a01-6480fb634889>