Include a mention of --cbr in the repo docs.
I've been using repo for the first time recently after a change of team
into the Android org, and one of the things I found myself getting
frustrated at was the need to get dropped into my editor to select the
brance I want to upload after an invocation of "repo upload".

The online docs make no mention of the --cbr flag, which removes my
frustration. This CL aims to fix that.

This is my first contribution to the devsite documentation, and I have
no idea how to preview the change I've made. I'd appreciate some
guidance on that.

Test: none, not sure how, halp?
Change-Id: I5109b84027339523cb79e5e6f40864dd1f8651c5